requirements.txt
colorama==0.4.6
opencv-python>=4.8.0
moviepy>=1.0.3
google-generativeai>=0.3.0
openai>=1.0.0
requests>=2.31.0
pillow>=10.0.0
pydub>=0.25.1
